corection inharitance User & new file
