+++ CarRental/Models/Car.cs
@@ -23,6 +23,3 @@ internal class Car
 
 }
 
-
-
-
